UIE-190 Auth Fetch Breakout part-1  - first wave of larger cleanup.  - remove bulk of circular dependencies in auth.ts and ajax-common.ts death-loop by breaking out specialized app-load bits in auth.ts
