+++ src/auth/oidc-broker.ts
@@ -1,7 +1,5 @@
-import _ from 'lodash/fp';
 import { ExtraSigninRequestArgs, IdTokenClaims, User, UserManager, WebStorageStateStore } from 'oidc-client-ts';
 import { AuthContextProps } from 'react-oidc-context';
-import { OAuth2, OidcConfig } from 'src/libs/ajax/OAuth2';
 import { getLocalStorage } from 'src/libs/browser-storage';
 import { getConfig } from 'src/libs/config';
 import { oidcStore } from 'src/libs/state';
@@ -46,12 +44,6 @@ export const getOidcConfig = () => {
     redirect_uri: '', // this field is not being used currently, but is expected from UserManager
   };
 };
-
-// This is the first thing that happens on app load.
-export const initializeClientId = _.memoize(async (): Promise<void> => {
-  const oidcConfig: OidcConfig = await OAuth2().getConfiguration();
-  oidcStore.update((state) => ({ ...state, config: oidcConfig }));
-});
 
 export const initializeOidcUserManager = () => {
   const userManager: UserManager = new UserManager(getOidcConfig());
